Add files via upload
Adding CMIP7_volcanic_aerosol_wl_interpolater-midpoint.py

This is a general purpose tool that is being made available for climate model users. It simply performs the task of interpolating CMIP7 source data of aerosol optical properties from the wavelengths specified in the file to wavelengths that the users can specify (overwrite) in a working copy of this python script.